refactor: remove redundant comments
