+++ react-rs/src/components/Gallery/Gallery.module.scss
