+++ react-rs/src/components/Gallery/Gallery.tsx
