+++ react-rs/src/components/Gallery/ImageCard/ImageCard.module.scss
