+++ react-rs/src/components/Gallery/Skeleton/Skeleton.module.scss
