+++ react-rs/src/components/Gallery/Skeleton/Skeleton.tsx
